+++ tests/helpers/testers.py
@@ -121,6 +121,7 @@ def _class_test(
     device: str = "cpu",
     fragment_kwargs: bool = False,
     check_scriptable: bool = True,
+    check_state_dict: bool = True,
     **kwargs_update: Any,
 ):
     """Utility function doing the actual comparison between class metric and reference metric.
@@ -151,9 +152,7 @@ def _class_test(
         metric_args = {}
 
     # Instantiate metric
-    metric = metric_class(
-        compute_on_step=check_dist_sync_on_step or check_batch, dist_sync_on_step=dist_sync_on_step, **metric_args
-    )
+    metric = metric_class(dist_sync_on_step=dist_sync_on_step, **metric_args)
     with pytest.raises(RuntimeError):
         metric.is_differentiable = not metric.is_differentiable
     with pytest.raises(RuntimeError):
@@ -180,7 +179,7 @@ def _class_test(
         batch_result = metric(preds[i], target[i], **batch_kwargs_update)
 
         if metric.dist_sync_on_step and check_dist_sync_on_step and rank == 0:
-            if isinstance(preds, torch.Tensor):
+            if isinstance(preds, Tensor):
                 ddp_preds = torch.cat([preds[i + r] for r in range(worldsize)]).cpu()
                 ddp_target = torch.cat([target[i + r] for r in range(worldsize)]).cpu()
             else:
@@ -203,8 +202,8 @@ def _class_test(
                 k: v.cpu() if isinstance(v, Tensor) else v
                 for k, v in (batch_kwargs_update if fragment_kwargs else kwargs_update).items()
             }
-            preds_ = preds[i].cpu() if isinstance(preds, torch.Tensor) else preds[i]
-            target_ = target[i].cpu() if isinstance(target, torch.Tensor) else target[i]
+            preds_ = preds[i].cpu() if isinstance(preds, Tensor) else preds[i]
+            target_ = target[i].cpu() if isinstance(target, Tensor) else target[i]
             sk_batch_result = sk_metric(preds_, target_, **batch_kwargs_update)
             if isinstance(batch_result, dict):
                 for key in batch_result.keys():
@@ -215,6 +214,10 @@ def _class_test(
     # check that metrics are hashable
     assert hash(metric)
 
+    # assert that state dict is empty
+    if check_state_dict:
+        assert metric.state_dict() == {}
+
     # check on all batches on all ranks
     result = metric.compute()
     if isinstance(result, dict):
@@ -223,7 +226,7 @@ def _class_test(
     else:
         _assert_tensor(result)
 
-    if isinstance(preds, torch.Tensor):
+    if isinstance(preds, Tensor):
         total_preds = torch.cat([preds[i] for i in range(num_batches)]).cpu()
         total_target = torch.cat([target[i] for i in range(num_batches)]).cpu()
     else:
@@ -569,7 +572,7 @@ class DummyMetric(Metric):
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
-        self.add_state("x", tensor(0.0), dist_reduce_fx=None)
+        self.add_state("x", tensor(0.0), dist_reduce_fx="sum")
 
     def update(self):
         pass
@@ -583,7 +586,7 @@ class DummyListMetric(Metric):
 
     def __init__(self):
         super().__init__()
-        self.add_state("x", [], dist_reduce_fx=None)
+        self.add_state("x", [], dist_reduce_fx="cat")
 
     def update(self):
         pass