+++ torchmetrics/functional/__init__.py
@@ -21,7 +21,7 @@ from torchmetrics.functional.classification.average_precision import average_pre
 from torchmetrics.functional.classification.calibration_error import calibration_error
 from torchmetrics.functional.classification.cohen_kappa import cohen_kappa
 from torchmetrics.functional.classification.confusion_matrix import confusion_matrix
-from torchmetrics.functional.classification.dice import dice_score
+from torchmetrics.functional.classification.dice import dice, dice_score
 from torchmetrics.functional.classification.f_beta import f1_score, fbeta_score
 from torchmetrics.functional.classification.hamming import hamming_distance
 from torchmetrics.functional.classification.hinge import hinge_loss
@@ -69,6 +69,7 @@ from torchmetrics.functional.retrieval.fall_out import retrieval_fall_out
 from torchmetrics.functional.retrieval.hit_rate import retrieval_hit_rate
 from torchmetrics.functional.retrieval.ndcg import retrieval_normalized_dcg
 from torchmetrics.functional.retrieval.precision import retrieval_precision
+from torchmetrics.functional.retrieval.precision_recall_curve import retrieval_precision_recall_curve
 from torchmetrics.functional.retrieval.r_precision import retrieval_r_precision
 from torchmetrics.functional.retrieval.recall import retrieval_recall
 from torchmetrics.functional.retrieval.reciprocal_rank import retrieval_reciprocal_rank
@@ -106,6 +107,7 @@ __all__ = [
     "coverage_error",
     "tweedie_deviance_score",
     "dice_score",
+    "dice",
     "error_relative_global_dimensionless_synthesis",
     "explained_variance",
     "extended_edit_distance",
@@ -146,6 +148,7 @@ __all__ = [
     "retrieval_r_precision",
     "retrieval_recall",
     "retrieval_reciprocal_rank",
+    "retrieval_precision_recall_curve",
     "roc",
     "rouge_score",
     "sacre_bleu_score",