+++ torchmetrics/functional/text/bert.py
@@ -18,6 +18,7 @@ from warnings import warn
 
 import torch
 from torch import Tensor
+from torch.nn import Module
 from torch.utils.data import DataLoader
 
 from torchmetrics.functional.text.helper_embedding_metric import (
@@ -44,13 +45,13 @@ _DEFAULT_MODEL = "roberta-large"
 def _get_embeddings_and_idf_scale(
     dataloader: DataLoader,
     target_len: int,
-    model: torch.nn.Module,
+    model: Module,
     device: Optional[Union[str, torch.device]] = None,
     num_layers: Optional[int] = None,
     all_layers: bool = False,
     idf: bool = False,
     verbose: bool = False,
-    user_forward_fn: Callable[[torch.nn.Module, Dict[str, Tensor]], Tensor] = None,
+    user_forward_fn: Callable[[Module, Dict[str, Tensor]], Tensor] = None,
 ) -> Tuple[Tensor, Tensor]:
     """Calculate sentence embeddings and the inverse-document-frequency scaling factor.
     Args:
@@ -69,7 +70,7 @@ def _get_embeddings_and_idf_scale(
             ``torch.Tensor``.
 
     Return:
-        A tuple of torch.Tensors containing the model's embeddings and the normalized tokens IDF.
+        A tuple of ``torch.Tensor``s containing the model's embeddings and the normalized tokens IDF.
         When ``idf = False``, tokens IDF is not calculated, and a matrix of mean weights is returned instead.
         For a single sentence, ``mean_weight = 1/seq_len``, where ``seq_len`` is a sum over the corresponding
         ``attention_mask``.
@@ -236,9 +237,9 @@ def bert_score(
     model_name_or_path: Optional[str] = None,
     num_layers: Optional[int] = None,
     all_layers: bool = False,
-    model: Optional[torch.nn.Module] = None,
+    model: Optional[Module] = None,
     user_tokenizer: Any = None,
-    user_forward_fn: Callable[[torch.nn.Module, Dict[str, Tensor]], Tensor] = None,
+    user_forward_fn: Callable[[Module, Dict[str, Tensor]], Tensor] = None,
     verbose: bool = False,
     idf: bool = False,
     device: Optional[Union[str, torch.device]] = None,