+++ torchmetrics/retrieval/precision.py
@@ -48,12 +48,6 @@ class RetrievalPrecision(RetrievalMetric):
             Ignore predictions where the target is equal to this number.
         k: consider only the top k elements for each query (default: ``None``, which considers them all)
         adaptive_k: adjust ``k`` to ``min(k, number of documents)`` for each query
-        compute_on_step:
-            Forward only calls ``update()`` and returns None if this is set to False.
-
-            .. deprecated:: v0.8
-                Argument has no use anymore and will be removed v0.9.
-
         kwargs: Additional keyword arguments, see :ref:`Metric kwargs` for more info.
 
     Raises:
@@ -76,7 +70,9 @@ class RetrievalPrecision(RetrievalMetric):
         tensor(0.5000)
     """
 
-    higher_is_better = True
+    is_differentiable: bool = False
+    higher_is_better: bool = True
+    full_state_update: bool = False
 
     def __init__(
         self,
@@ -84,13 +80,11 @@ class RetrievalPrecision(RetrievalMetric):
         ignore_index: Optional[int] = None,
         k: Optional[int] = None,
         adaptive_k: bool = False,
-        compute_on_step: Optional[bool] = None,
         **kwargs: Dict[str, Any],
     ) -> None:
         super().__init__(
             empty_target_action=empty_target_action,
             ignore_index=ignore_index,
-            compute_on_step=compute_on_step,
             **kwargs,
         )