+++ torchmetrics/retrieval/reciprocal_rank.py
@@ -43,12 +43,6 @@ class RetrievalMRR(RetrievalMetric):
             - ``'error'``: raise a ``ValueError``
 
         ignore_index: Ignore predictions where the target is equal to this number.
-        compute_on_step:
-            Forward only calls ``update()`` and returns None if this is set to False.
-
-            .. deprecated:: v0.8
-                Argument has no use anymore and will be removed v0.9.
-
         kwargs: Additional keyword arguments, see :ref:`Metric kwargs` for more info.
 
     Raises:
@@ -67,7 +61,9 @@ class RetrievalMRR(RetrievalMetric):
         tensor(0.7500)
     """
 
-    higher_is_better = True
+    is_differentiable: bool = False
+    higher_is_better: bool = True
+    full_state_update: bool = False
 
     def _metric(self, preds: Tensor, target: Tensor) -> Tensor:
         return retrieval_reciprocal_rank(preds, target)