+++ torchmetrics/text/bert.py
@@ -16,6 +16,7 @@ from warnings import warn
 
 import torch
 from torch import Tensor
+from torch.nn import Module
 
 from torchmetrics.functional.text.bert import bert_score
 from torchmetrics.functional.text.helper_embedding_metric import _preprocess_text
@@ -66,7 +67,7 @@ class BERTScore(Metric):
             A user's own forward function used in a combination with `user_model`. This function must take `user_model`
             and a python dictionary of containing `"input_ids"` and `"attention_mask"` represented by `torch.Tensor`
             as an input and return the model's output represented by the single `torch.Tensor`.
-        verbose: An indication of whether a progress bar to be displayed during the embeddings calculation.
+        verbose: An indication of whether a progress bar to be displayed during the embeddings' calculation.
         idf: An indication whether normalization using inverse document frequencies should be used.
         device: A device to be used for calculation.
         max_length: A maximum length of input sequences. Sequences longer than `max_length` are to be trimmed.
@@ -82,12 +83,6 @@ class BERTScore(Metric):
             of the files from `BERT_score`_.
         baseline_path: A path to the user's own local csv/tsv file with the baseline scale.
         baseline_url: A url path to the user's own  csv/tsv file with the baseline scale.
-        compute_on_step:
-            Forward only calls ``update()`` and returns None if this is set to False.
-
-            .. deprecated:: v0.8
-                Argument has no use anymore and will be removed v0.9.
-
         kwargs: Additional keyword arguments, see :ref:`Metric kwargs` for more info.
 
     Returns:
@@ -105,8 +100,10 @@ class BERTScore(Metric):
         {'f1': [1.0, 0.996], 'precision': [1.0, 0.996], 'recall': [1.0, 0.996]}
     """
 
-    is_differentiable = False
-    higher_is_better = True
+    is_differentiable: bool = False
+    higher_is_better: bool = True
+    full_state_update: bool = False
+
     preds_input_ids: List[Tensor]
     preds_attention_mask: List[Tensor]
     target_input_ids: List[Tensor]
@@ -117,9 +114,9 @@ class BERTScore(Metric):
         model_name_or_path: Optional[str] = None,
         num_layers: Optional[int] = None,
         all_layers: bool = False,
-        model: Optional[torch.nn.Module] = None,
+        model: Optional[Module] = None,
         user_tokenizer: Optional[Any] = None,
-        user_forward_fn: Callable[[torch.nn.Module, Dict[str, torch.Tensor]], torch.Tensor] = None,
+        user_forward_fn: Callable[[Module, Dict[str, Tensor]], Tensor] = None,
         verbose: bool = False,
         idf: bool = False,
         device: Optional[Union[str, torch.device]] = None,
@@ -131,10 +128,9 @@ class BERTScore(Metric):
         rescale_with_baseline: bool = False,
         baseline_path: Optional[str] = None,
         baseline_url: Optional[str] = None,
-        compute_on_step: Optional[bool] = None,
         **kwargs: Dict[str, Any],
     ):
-        super().__init__(compute_on_step=compute_on_step, **kwargs)
+        super().__init__(**kwargs)
         self.model_name_or_path = model_name_or_path or _DEFAULT_MODEL
         self.num_layers = num_layers
         self.all_layers = all_layers
@@ -151,8 +147,8 @@ class BERTScore(Metric):
         self.rescale_with_baseline = rescale_with_baseline
         self.baseline_path = baseline_path
         self.baseline_url = baseline_url
-        self.preds: Dict[str, List[torch.Tensor]] = {"input_ids": [], "attention_mask": []}
-        self.target: Dict[str, List[torch.Tensor]] = {"input_ids": [], "attention_mask": []}
+        self.preds: Dict[str, List[Tensor]] = {"input_ids": [], "attention_mask": []}
+        self.target: Dict[str, List[Tensor]] = {"input_ids": [], "attention_mask": []}
 
         if user_tokenizer:
             self.tokenizer = user_tokenizer